% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopeAdductMetaboliteAnnotation.R
\name{adductAnnotation}
\alias{adductAnnotation}
\title{adductAnnotation}
\usage{
adductAnnotation(id = NA, formula = "C6H14N4O2", adduct = "M+H",
  polarity = c("positive", "negative"), mz = 175.118, rt = 961.6225,
  adduct.table, peak.mz, peak.rt, cor, mz.tol = 25, rt.tol = 3,
  cor.tol = 0.5)
}
\arguments{
\item{id}{The kegg ID of metabolite.}

\item{formula}{The formula of metabolite.}

\item{adduct}{The adduct of metabolite.}

\item{polarity}{The mode of data.}

\item{mz}{The mz of metabolite.}

\item{rt}{The rt of metabolite.}

\item{adduct.table}{The adduct table.}

\item{peak.mz}{The mz of peaks.}

\item{peak.rt}{The rt of peaks.}

\item{cor}{The correlations between metabolite and peaks.}

\item{mz.tol}{The mz tol.}

\item{rt.tol}{The rt tol (second).}

\item{cor.tol}{The correlation tol.}
}
\value{
Isotope annotation information of peaks.
}
\description{
adduct annotation
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
