% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisGeneration.R
\name{analysisGeneration}
\alias{analysisGeneration}
\title{analysisGeneration}
\usage{
analysisGeneration(polarity = c("positive", "negative", "both"),
  pos.path = ".", neg.path = ".", data.pos = "data.csv",
  data.neg = "data.csv", sample.info = "sample.info.csv",
  output.path = ".", correct = TRUE, p.cutoff = 0.01,
  candidate.num = candidate.num, type = c("html", "pdf", "all"))
}
\arguments{
\item{polarity}{The polarity}

\item{pos.path}{The directory of positive data.}

\item{neg.path}{The directory of negative data.}

\item{data.pos}{The name of positive data.}

\item{data.neg}{The name of negative data.}

\item{sample.info}{The name of sample.info.}

\item{output.path}{Output directory.}

\item{correct}{Correct p value or not.}

\item{p.cutoff}{P value cutoff.}

\item{candidate.num}{Top candidate.num metabolites for each peak.}

\item{type}{Analysis report type.}
}
\value{
Analysis report.
}
\description{
Generate analysis report
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
