% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{checkElement}
\alias{checkElement}
\title{checkElement}
\usage{
checkElement(formula = "C9H11NO2", adduct = "M-H2O+H")
}
\arguments{
\item{formula}{The formula of metabolite.}

\item{adduct}{The adduct.}
}
\value{
valid, return TRUE; invalid.
}
\description{
Check a formula can add one adduct or not.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
