% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{compareAnnotation}
\alias{compareAnnotation}
\title{compareAnnotation}
\usage{
compareAnnotation(tags2_1, tags2_2)
}
\arguments{
\item{tags2_1}{The tags2 data.}

\item{tags2_2}{The tags2 data.}
}
\value{
The comparesion result.
}
\description{
compareAnnotation.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
