% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakClass.R
\name{equalPeakinfo}
\alias{equalPeakinfo}
\title{equalPeakinfo}
\usage{
equalPeakinfo(object1, object2)
}
\arguments{
\item{object1}{peakInfo.}

\item{object2}{peakInfo.}
}
\value{
TRUE or FALSE.
}
\description{
Compare two peakInfo data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
