% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakClass.R
\name{equalPeakinfo1}
\alias{equalPeakinfo1}
\title{equalPeakinfo1}
\usage{
equalPeakinfo1(x, y)
}
\arguments{
\item{x}{The tags2 data.}

\item{y}{The tags2 data.}
}
\value{
TRUE or FALSE.
}
\description{
Compare two tags2 data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
