% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyFeature.R
\name{GetDiffMZppm}
\alias{GetDiffMZppm}
\title{GetDiffMZppm}
\usage{
GetDiffMZppm(mz, mz.ppm.thr = NULL)
}
\arguments{
\item{mz}{mz}

\item{mz.ppm.thr}{mz.ppm.thr}
}
\description{
GetDiffMZppm
}
\author{
Hao Li, Yandong Yin, Kai Weng
\email{shenxt@sioc.ac.cn}
}
