% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyFeature.R
\name{GetDotProduct}
\alias{GetDotProduct}
\title{GetDotProduct}
\usage{
GetDotProduct(x, y)
}
\arguments{
\item{x}{x}

\item{y}{y}
}
\description{
GetDotProduct
}
\author{
Hao Li, Yandong Yin, Kai Weng
\email{shenxt@sioc.ac.cn}
}
