% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyFeature.R
\name{GetMatchResult}
\alias{GetMatchResult}
\title{GetMatchResult}
\usage{
GetMatchResult(spec.exp, spec.lib, weight.int = 1, weight.mz = 0,
  ppm.ms2match = 30, mz.ppm.thr = 400, ppm.sanity.check = 100,
  is.sanity.check = FALSE, direction = direction, ...)
}
\arguments{
\item{spec.exp}{spec.exp}

\item{spec.lib}{spec.lib}

\item{weight.int}{weight.int}

\item{weight.mz}{weight.mz}

\item{ppm.ms2match}{ppm.ms2match}

\item{mz.ppm.thr}{mz.ppm.thr}

\item{ppm.sanity.check}{ppm.sanity.check}

\item{is.sanity.check}{is.sanity.check}

\item{direction}{direction}

\item{...}{other parameters}
}
\description{
GetMatchResult
}
\author{
Hao Li, Yandong Yin, Kai Weng
\email{shenxt@sioc.ac.cn}
}
