% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getNeighbor}
\alias{getNeighbor}
\title{getNeighbor}
\usage{
getNeighbor(metabolite.id, step = 1, graph)
}
\arguments{
\item{metabolite.id}{The metabolite ID.}

\item{step}{The reaction step.}

\item{graph}{The kegg.rpair2.}
}
\value{
The neighbor information.
}
\description{
Get the neighbors of one metabolite in graph.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
