% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyFeature.R
\name{GetSpec2Match}
\alias{GetSpec2Match}
\title{GetSpec2Match}
\usage{
GetSpec2Match(spec.exp, spec.lib, ppm.ms2match = 30, mz.ppm.thr = 400,
  direction = c("reverse", "forward"))
}
\arguments{
\item{spec.exp}{spec.exp}

\item{spec.lib}{spec.lib}

\item{ppm.ms2match}{ppm.ms2match}

\item{mz.ppm.thr}{mz.ppm.thr}

\item{direction}{direction}
}
\description{
GetSpec2Match
}
\author{
Hao Li, Yandong Yin, Kai Weng
\email{shenxt@sioc.ac.cn}
}
