% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyFeature.R
\name{GetWeightedInt}
\alias{GetWeightedInt}
\title{GetWeightedInt}
\usage{
GetWeightedInt(spec, weight.mz = 0, weight.int = 1)
}
\arguments{
\item{spec}{spec}

\item{weight.mz}{weight.mz}

\item{weight.int}{weight.int}
}
\description{
GetWeightedInt
}
\author{
Hao Li, Yandong Yin, Kai Weng
\email{shenxt@sioc.ac.cn}
}
