% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalysis.R
\name{gseAnalysis}
\alias{gseAnalysis}
\title{gseAnalysis}
\usage{
gseAnalysis(gene.id, species = c("hsa", "dme", "mmu", "rat", "bta", "gga",
  "dre", "cel", "sce", "ath", "smm", "pfa", "tbr", "eco", "ppu", "syf"),
  test.method = c("hypergeometric", "fisher"))
}
\arguments{
\item{gene.id}{The vector of gene IDs.}

\item{species}{The species. "hsa" is Homo sapiens (human),
"dme" is Drosophlia melanogaster (fruit fly),
"mmu" is Mus musculus (mouse), "rat" is Rattus norvegicus (rat),
"bta" is Bos taurus (cow), "gga" is Gallus domesticus (chicken),
"dre" is Danio rerio (zebrafish), "cel" is Caenorharomyces elegans (nematode),
 "sce" is Saccharomyces cerevisaiae (yeast), "ath" is Arabidopsis thaliana (thale cress),
 "smm" is Schistosoma mansoni, "pfa" is Plasmodum falciparum 3D7,
 "tbr" is Trypanosoma brucei, "eco" is Escherichia coli K-12 MG1655,
"ppu" is Pseudomonas putida KT2440, "syf" is Synechococcus elongatus.}

\item{test.method}{Hypergeometric or fisher test.}
}
\value{
The MSE analysis result.
}
\description{
Gene sets enrichment analysis.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
