% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyFeature.R
\name{IdentifyFeature}
\alias{IdentifyFeature}
\title{IdentifyFeature}
\usage{
IdentifyFeature(spec.exp, spec.lib, is.include.precursor = FALSE,
  is.tune.ms2.lib = FALSE, direction = "forward", ...)
}
\arguments{
\item{spec.exp}{spec.exp.}

\item{spec.lib}{spec.lib.}

\item{is.include.precursor}{is.include.precursor.}

\item{is.tune.ms2.lib}{is.tune.ms2.lib.}

\item{direction}{direction.}

\item{...}{other parameters}
}
\description{
IdentifyFeature
}
\author{
Hao Li, Yandong Yin, Kai Weng
\email{shenxt@sioc.ac.cn}
}
