% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyFeature.R
\name{MatchFromTemp}
\alias{MatchFromTemp}
\title{MatchFromTemp}
\usage{
MatchFromTemp(spec, temp)
}
\arguments{
\item{spec}{spec}

\item{temp}{temp}
}
\description{
MatchFromTemp
}
\author{
Hao Li, Yandong Yin, Kai Weng
\email{shenxt@sioc.ac.cn}
}
