% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyFeature.R
\name{MatchSpec}
\alias{MatchSpec}
\title{MatchSpec}
\usage{
MatchSpec(spec, ppm.ms2match = 30, mz.ppm.thr = 400)
}
\arguments{
\item{spec}{spec}

\item{ppm.ms2match}{ppm.ms2match}

\item{mz.ppm.thr}{mz.ppm.thr}
}
\description{
MatchSpec
}
\author{
Hao Li, Yandong Yin, Kai Weng
\email{shenxt@sioc.ac.cn}
}
