% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{memoryList}
\alias{memoryList}
\title{memoryList}
\usage{
memoryList(decreasing = TRUE, n = 10)
}
\arguments{
\item{decreasing}{Logical. List the object from big to small or not.}

\item{n}{List how many objects.}
}
\value{
The size of objects.
}
\description{
List the size of objects.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
