% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metABM.R
\name{metABM}
\alias{metABM}
\title{metABM}
\usage{
metABM(annotation.result = "ms2.match.annotation.result.csv",
  ms2.data = "ms2", prefer.adduct = c("all", "M+H", "M+Na", "M-H"),
  use.default.md = TRUE, column = c("hilic", "rp"),
  polarity = c("positive", "negative"), threads = 3, path = file.path(".",
  "MS2_match_result"), output.path = file.path(".", "MRN_annotation_result"),
  max.isotope = 4, mz.tol = 25, rt.tol1 = 3, rt.tol2 = 30,
  cor.tol = 0, int.tol = 500, dp.tol = 0.5, max.step = 3,
  score.cutoff = 0, remain = FALSE, remain.per = 0.5,
  seed.neighbor.match.plot = FALSE, candidate.num = 3)
}
\arguments{
\item{annotation.result}{The name of annotation result from ms2Annotation.}

\item{ms2.data}{The name of ms2 data from ms2Annotation.}

\item{prefer.adduct}{Which adduct you want to use for RT prediction.}

\item{use.default.md}{Use default molecular descriptors for RT prediction or not.}

\item{column}{hilic or rp.}

\item{polarity}{positive or negative.}

\item{threads}{The number of threads.}

\item{path}{The work directory.}

\item{output.path}{The directory for output results.}

\item{max.isotope}{The number of isotope peaks}

\item{mz.tol}{The m/z tolerance. Default is 25 ppm.}

\item{rt.tol1}{The RT tolerance for isotope and adduct annotation. (second)
Default is 3.}

\item{rt.tol2}{The RT tolerance for metabolite annotation. (\%) Default
is 30\%.}

\item{cor.tol}{The correlation tolerance.}

\item{int.tol}{The intensity ratio tolerance of isotope annotation.}

\item{dp.tol}{The tolerance of dot product.}

\item{max.step}{The max number of reaction step.}

\item{score.cutoff}{Score cutoff of annotations.}

\item{remain}{Remain some seeds as validation or not.}

\item{remain.per}{The percentage of remained seeds. Default is 50\%.}

\item{seed.neighbor.match.plot}{Output MS/MS match plot of seed and neighbor or not.}

\item{candidate.num}{How many candidates for peaks are outputted. Default is 3.}
}
\value{
Annotation result.
}
\description{
Annotate peaks based on metabolic reaction network (MRN).
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
