% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopeAdductMetaboliteAnnotation.R
\name{metAnnotation}
\alias{metAnnotation}
\title{metAnnotation}
\usage{
metAnnotation(metabolite.name = "M175T962", metabolite.id = "C00062",
  formula = "C6H14N4O2", adduct = "M+H", polarity = c("positive",
  "negative"), mz = 175.118, rt = 961.6225, peak.mz, peak.rt, cor, ms2,
  mz.tol = 25, rt.tol = 30, cor.tol = 0, dp.tol = 0.5, step = 1,
  metabolite, metabolic.network, adduct.table)
}
\arguments{
\item{metabolite.name}{The metabolite name.}

\item{metabolite.id}{The metabolite ID.}

\item{formula}{The formula of metabolite.}

\item{adduct}{The adduct of metabolite.}

\item{polarity}{The polarity.}

\item{mz}{The mz of metabolite.}

\item{rt}{The RT of metabolite.}

\item{peak.mz}{The mz of all peaks.}

\item{peak.rt}{The RT of all peaks.}

\item{cor}{The correlation of metabolite between all peaks.}

\item{ms2}{The ms2 data of peak table.}

\item{mz.tol}{The mz tolerance.}

\item{rt.tol}{The RT tolerance for metabolite annotation. (\%)}

\item{cor.tol}{The cor tolerance.}

\item{dp.tol}{The tolerance of dot product.}

\item{step}{The reaction step.}

\item{metabolite}{The kegg compound database.}

\item{metabolic.network}{kegg.rpair2}

\item{adduct.table}{Adduct table.}
}
\value{
The metabolite annotation information.
}
\description{
Annotate peak table from one metabolite.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
