% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metEnrichment.R
\name{metEnrichment}
\alias{metEnrichment}
\title{metEnrichment}
\usage{
metEnrichment(annotation.table, sample.pos, sample.neg, group,
  column = c("hilic", "rp"), candidate.num = 5, sample.info = sample.info,
  polarity = c("positive", "negative", "both"), species = c("hsa", "dme",
  "mmu", "rat", "bta", "gga", "dre", "cel", "sce", "ath", "smm", "pfa", "tbr",
  "eco", "ppu", "syf"), pos.path = pos.path, neg.path = neg.path,
  output.path = output.path)
}
\arguments{
\item{annotation.table}{The metabolite annotation result which used for
pathway enrichment analysis. The data frame should contain two column,
the column names are peak.name and KEGG.ID.}

\item{sample.pos}{Positive annotation.result from ms2Annotation.}

\item{sample.neg}{Negative annotation.result from ms2Annotation.}

\item{group}{The group you want to use.}

\item{column}{Column, "hilic" or "rp".}

\item{candidate.num}{Top candidate.num metabolites for each peak.}

\item{sample.info}{Sample information.}

\item{polarity}{The polarity mode of data. "positive", "negative" or "both".}

\item{species}{The species. "hsa" is Homo sapiens (human),
"dme" is Drosophlia melanogaster (fruit fly),
"mmu" is Mus musculus (mouse), "rat" is Rattus norvegicus (rat),
"bta" is Bos taurus (cow), "gga" is Gallus domesticus (chicken),
"dre" is Danio rerio (zebrafish), "cel" is Caenorharomyces elegans (nematode),
 "sce" is Saccharomyces cerevisaiae (yeast), "ath" is Arabidopsis thaliana (thale cress),
 "smm" is Schistosoma mansoni, "pfa" is Plasmodum falciparum 3D7,
 "tbr" is Trypanosoma brucei, "eco" is Escherichia coli K-12 MG1655,
"ppu" is Pseudomonas putida KT2440, "syf" is Synechococcus elongatus.}

\item{pos.path}{Directory of positive results.}

\item{neg.path}{Directory of negative results.}

\item{output.path}{The directory to output results.}
}
\description{
Metabolic enrichment analysis.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
