% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metModule.R
\name{metModule}
\alias{metModule}
\title{metModule}
\usage{
metModule(annotation.result.pos, annotation.result.neg, tags2.pos = tags2.pos,
  tags2.neg = tags2.neg, p.value.pos = p.value.pos,
  p.value.neg = p.value.neg, foldchange.pos = foldchange.pos,
  foldchange.neg = foldchange.neg, rt.result = rt.result,
  pos.path = pos.path, neg.path = neg.path, sample.info = sample.info,
  polarity = c("positive", "negative", "both"), group, max.isotope = 4,
  uni.test = c("t", "wilcox", "anova"), column = c("hilic", "rp"),
  output.path = ".", correct = TRUE, p.cutoff = 0.01, mz.tol = 25,
  rt.tol1 = 3, rt.tol2 = 30, cor.tol = 0, int.tol = 500, threads = 3,
  use.old.null = FALSE, species = c("hsa", "dme", "mmu", "rat", "bta",
  "gga", "dre", "cel", "sce", "ath", "smm", "pfa", "tbr", "eco", "ppu", "syf"),
  use.all.kegg.id = FALSE, only.mrn.annotation = FALSE,
  output.module.information = TRUE)
}
\arguments{
\item{annotation.result.pos}{Positive annotation.result from ms2Annotation.}

\item{annotation.result.neg}{Negative annotation.result from ms2Annotation.}

\item{tags2.pos}{Tags2 result of positive data.}

\item{tags2.neg}{Tags2 result of negative data.}

\item{p.value.pos}{Positive p values.}

\item{p.value.neg}{Negative p values.}

\item{foldchange.pos}{Positive fold change.}

\item{foldchange.neg}{Negative fold change.}

\item{rt.result}{Predicted RT results from rtPrediction.}

\item{pos.path}{Directory of positive results.}

\item{neg.path}{Directory of negative results.}

\item{sample.info}{Sample information.}

\item{polarity}{"positive", "negative" or "both".}

\item{group}{The group you want to use.}

\item{max.isotope}{The max number of isotope.}

\item{uni.test}{The method of univariate test.}

\item{column}{hilic or rp.}

\item{output.path}{The directory to output results.}

\item{correct}{Correct p value or not.}

\item{p.cutoff}{The cutoff of p value.}

\item{mz.tol}{The mz tolerance of KEGG matching.}

\item{rt.tol1}{The RT tolerance of isotope annotation (second).}

\item{rt.tol2}{The RT tolerance of KEGG matching (\%).}

\item{cor.tol}{The tolerance of correlation.}

\item{int.tol}{The tolerance of intensity ratio (\%).}

\item{threads}{The number of threads.}

\item{use.old.null}{Use old null distribution of not.}

\item{species}{The species. "hsa" is Homo sapiens (human),
"dme" is Drosophlia melanogaster (fruit fly),
"mmu" is Mus musculus (mouse), "rat" is Rattus norvegicus (rat),
"bta" is Bos taurus (cow), "gga" is Gallus domesticus (chicken),
"dre" is Danio rerio (zebrafish), "cel" is Caenorharomyces elegans (nematode),
 "sce" is Saccharomyces cerevisaiae (yeast), "ath" is Arabidopsis thaliana (thale cress),
 "smm" is Schistosoma mansoni, "pfa" is Plasmodum falciparum 3D7,
 "tbr" is Trypanosoma brucei, "eco" is Escherichia coli K-12 MG1655,
"ppu" is Pseudomonas putida KT2440, "syf" is Synechococcus elongatus.}

\item{use.all.kegg.id}{Use all annotations from KEGG database.
Default is FALSE.}

\item{only.mrn.annotation}{Only use MEN annotations or not.}

\item{output.module.information}{Output module information ot not.}
}
\description{
Metabolic module analysis.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
