% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metPathway.R
\name{metPathway}
\alias{metPathway}
\title{metPathway}
\usage{
metPathway(annotation.result = "ms2.match.annotation.result.csv", group,
  max.isotope = 4, uni.test = c("t", "wilcox", "anova"),
  column = c("hilic", "rp"), polarity = c("positive", "negative"),
  path = paste(file.path(".", "MRN_annotation_result"), ifelse(polarity ==
  "positive", "POS", "NEG"), sep = "_"), output.path = paste(file.path(".",
  "Dysregulated_network_analysis_result"), ifelse(polarity == "positive", "POS",
  "NEG"), sep = "_"), correct = TRUE, p.cutoff = 0.01, mz.tol = 25,
  rt.tol1 = 3, rt.tol2 = 30, cor.tol = 0, int.tol = 500, threads = 3,
  species = c("hsa", "dme", "mmu", "rat", "bta", "gga", "dre", "cel", "sce",
  "ath", "smm", "pfa", "tbr", "eco", "ppu", "syf"), use.all.kegg.id = FALSE,
  only.mrn.annotation = FALSE, use.old.null = FALSE)
}
\arguments{
\item{annotation.result}{Annotation.result from ms2Annotation.}

\item{group}{The group you want to use.}

\item{max.isotope}{The max number of isotope.}

\item{uni.test}{The method of univariate test.}

\item{column}{hilic or rp.}

\item{polarity}{positive or negative.}

\item{path}{The work directory.}

\item{output.path}{The directory for outputting results.}

\item{correct}{Correct p value or not.}

\item{p.cutoff}{The cutoff of p value. Default is 0.01.}

\item{mz.tol}{The mz tolerance of KEGG matching. Defalut is 25 ppm.}

\item{rt.tol1}{The RT tolerance of isotope annotation (second). Default is
3 s.}

\item{rt.tol2}{The RT tolerance of KEGG matching (\%). Default is 30\%.}

\item{cor.tol}{THe tolerance of correlation.}

\item{int.tol}{The tolerance of intensity ratio for
isotope annotation(\%).}

\item{threads}{The number of threads.}

\item{species}{The species. "hsa" is Homo sapiens (human),
"dme" is Drosophlia melanogaster (fruit fly),
"mmu" is Mus musculus (mouse), "rat" is Rattus norvegicus (rat),
"bta" is Bos taurus (cow), "gga" is Gallus domesticus (chicken),
"dre" is Danio rerio (zebrafish), "cel" is Caenorharomyces elegans (nematode),
 "sce" is Saccharomyces cerevisaiae (yeast), "ath" is Arabidopsis thaliana (thale cress),
 "smm" is Schistosoma mansoni, "pfa" is Plasmodum falciparum 3D7,
 "tbr" is Trypanosoma brucei, "eco" is Escherichia coli K-12 MG1655,
"ppu" is Pseudomonas putida KT2440, "syf" is Synechococcus elongatus.}

\item{use.all.kegg.id}{Use all annotations from KEGG database.
Default is FALSE.}

\item{only.mrn.annotation}{Only use MEN annotations or not.}

\item{use.old.null}{Use old null distribution of not. Default is FALSE.}
}
\description{
Metabolic pathway analysis.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
