% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleClass.R
\name{moduleplot}
\alias{moduleplot}
\title{moduleplot}
\usage{
moduleplot(object, cex.label = 1, mar, n, main, show = c("p", "overlap"),
  ...)
}
\arguments{
\item{object}{moduleInfo object.}

\item{cex.label}{The font size of label.}

\item{mar}{margin size.}

\item{n}{The number of pathway.}

\item{main}{Main of plot.}

\item{show}{p.value or voerlap.}

\item{...}{see ?barplot}
}
\description{
Barplot of pathway enrichment result.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
