% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Annotation.R
\name{ms2Annotation}
\alias{ms2Annotation}
\title{ms2Annotation}
\usage{
ms2Annotation(ms1.file = "data.csv", instrument = c("AgilentQTOF",
  "SciexTripleTOF", "OtherQTOF", "ThermoOrbitrap"), ms2.file,
  sample.info = "sample.info.csv", mz.tol = 25, rt.tol = 10,
  dp.cutoff = 0.8, polarity = c("positive", "negative"), path = ".",
  output.path = file.path(path, "MS2_match_result"), ms2.type = c("mgf",
  "mzXML", "msp"), column = c("hilic", "rp"), ce = c("30", "10", "20",
  "35,15", "40", "50"), ms2.match.plot = TRUE)
}
\arguments{
\item{ms1.file}{The name of ms1 peak table. Column 1 is "name", Column 2 is
"mz" and column is "rt".}

\item{instrument}{The instrument you used to acquire data. "AgilentQTOF",
"SciexTripleTOF", "OtherQTOF" and "ThermoOrbitrap" are supported.}

\item{ms2.file}{The vector of names of ms2 files. MS2 file must be mzXML.}

\item{sample.info}{The name of sample.info.}

\item{mz.tol}{mz tol for ms1 and ms2 data matching.}

\item{rt.tol}{RT tol for ms1 and ms2 data matching.}

\item{dp.cutoff}{The cutoff of dot product. Default is 0.8.}

\item{polarity}{The polarity of mode.}

\item{path}{Directory.}

\item{output.path}{The directory for outputing results.}

\item{ms2.type}{"mgf", "msp" or "mzXML".}

\item{column}{The column.}

\item{ce}{The collision energy.}

\item{ms2.match.plot}{Output MS2 match plot or not.}
}
\value{
Return the annotation result.
}
\description{
Annotate peak using MS/MS spectra in in-house database.
}
\author{
Xiaotao Shen, Yandong Yin
\email{shenxt@sioc.ac.cn}
}
