% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pasteElement}
\alias{pasteElement}
\title{pasteElement}
\usage{
pasteElement(formula = "C9H11NO2", element = "H", mode = c("plus",
  "minus"))
}
\arguments{
\item{formula}{The formula of metabolite.}

\item{element}{The element.}

\item{mode}{Add or remove a module}
}
\value{
A formula.
}
\description{
Paste formula and element.
Combine metabolite and adduct as a new sum formula.
If there are no enough element to remove, return NA.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
