% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnnotation.R
\name{readAnnotation}
\alias{readAnnotation}
\title{readAnnotation}
\usage{
readAnnotation(data = "data.csv", path = ".", rt.filter = FALSE,
  rt.tol = 30, inHouse.compound = inhouse.compound)
}
\arguments{
\item{data}{The file name (csv) from ms2Annotation}

\item{path}{The work directory}

\item{rt.filter}{Filter annotation according or not.}

\item{rt.tol}{The tolerance of RT (\%).}

\item{inHouse.compound}{in house compound database.}
}
\description{
Read the data from ms2Annotation and seperate it to
tags and sample
}
\author{
Xiaotao Shen
}
