% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{removeIsotope}
\alias{removeIsotope}
\title{removeIsotope}
\usage{
removeIsotope(result)
}
\arguments{
\item{result}{The MRN annotation table form MetDNA.}
}
\description{
remove isotope from result.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
