% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{sendMail}
\alias{sendMail}
\title{sendMail}
\usage{
sendMail(email.address = "zutaoshen@gmail.com", polarity = c("positive",
  "negative", "both"), from = "kagawabale@163.com",
  user.name = "kagawabale", pw = "19900521@2016", attach = TRUE)
}
\arguments{
\item{email.address}{Email address.}

\item{polarity}{Acquisition mode.}

\item{from}{The address of sender.}

\item{user.name}{User name.}

\item{pw}{Pass word.}

\item{attach}{Attach or not.}
}
\description{
Sent results to user using email.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
