% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{splitFormula}
\alias{splitFormula}
\title{splitFormula}
\usage{
splitFormula(formula = "C9H11NO2")
}
\arguments{
\item{formula}{The formula of metabolite.}
}
\value{
A splited formula.
}
\description{
Split a formula into element and number.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
