% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{sumFormula}
\alias{sumFormula}
\title{sumFormula}
\usage{
sumFormula(formula = "C9H11NO2", adduct = "M-H2O+H")
}
\arguments{
\item{formula}{The formula of metabolite.}

\item{adduct}{The adduct of metabolite.}
}
\value{
A sum formula.
}
\description{
Combine metabolite and adduct as a new sum formula.
If there are no enough element to remove, return NA.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
