% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tags2Match}
\alias{tags2Match}
\title{tags2Match}
\usage{
tags2Match(query = c("M102T561", "C00576"), tags2, is.seed = FALSE)
}
\arguments{
\item{query}{The ID of annotation.}

\item{tags2}{The tags2 data.}

\item{is.seed}{Is seed or not.}
}
\value{
The index of annotation.
}
\description{
Find annotation in tags2 data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
