% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{uniTest}
\alias{uniTest}
\title{uniTest}
\usage{
uniTest(sample, sample.info, uni.test = c("t", "wilcox", "anova"),
  alternative = c("two.sided", "less", "greater"), mu = 0, paired = FALSE,
  var.equal = FALSE, conf.level = 0.95, exact = NULL, correct = TRUE,
  conf.int = FALSE, threads = 3, ...)
}
\arguments{
\item{sample}{The sample abundance. Column is sample and row is peak.}

\item{sample.info}{The sample.information}

\item{uni.test}{Test method.}

\item{alternative}{see ?t.test.}

\item{mu}{see ?t.test.}

\item{paired}{see ?t.test.}

\item{var.equal}{see ?t.test.}

\item{conf.level}{see ?t.test.}

\item{exact}{see ?t.test.}

\item{correct}{see ?t.test.}

\item{conf.int}{see ?t.test.}

\item{threads}{How many threads do you want to use? Default is the number of
your PC threads - 3.}

\item{...}{See ?t.test.}
}
\description{
Univariate analysis.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
