% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{zipData}
\alias{zipData}
\title{zipData}
\usage{
zipData(polarity = c("positive", "negative", "both"))
}
\arguments{
\item{polarity}{Acquisition mode.}
}
\description{
Compress analysis result to zip.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
